requirements.txt
mysql-connector-python==8.3.0
requests==2.31.0
mysql-connector-python
requests
python-dotenv
